Update SyncCanvasAvatar.ps1
Removed code which exports student photographs from Synergy, as this is now handled by another independent script on a regular schedule. This was separated out because we have multiple systems which use student photos, so it makes more sense to have a separate folder full of images which is synced regularly, and have all systems read from there (including this one).